UI redesigned. Fixed the multiclicking while using the knob.

+++ include/ui/font_color_logic.h
@@ -7,10 +7,10 @@
 #define HANDLECOLORCHOICE_H
 
 
-void handleColorChoice(int delta, int& selectedColorID, uint16_t& activeButtonColor,
+void handleColorChoice(int8_t delta, int& selectedColorID, uint16_t& activeButtonColor,
     std::vector<uint16_t>& buttonColors,
     Window* activeWindow, Window* mainWindow, Window* settingsWindow);
 
-void handleFontChoice(int delta, int& selectedFontID, 
+void handleFontChoice(int8_t delta, int& selectedFontID, 
     std::vector<font_descriptor_t*>& availableFonts,Window* activeWindow,Window* mainWindow, Window* settingsWindow);
 #endif
